modify codes in Chapter1

+++ Chapter1/Chapter_1_unrun_code/1_27_Ackermann.cpp
@@ -6,9 +6,8 @@
 //
 
 #include <cfloat>
-#include <w32api/sqltypes.h>
 
-__int64 Ackermann(int m, int n){
+long long Ackermann(long long m, long long n){
     if (m == 0) return n+1;
     else if ( m > 0 && n == 0) return Ackermann( m - 1, 1);
     else if ( m > 0 && n > 0) return Ackermann(m - 1, Ackermann(m, n-1));
