+++ Chapter1/Chapter_1_unrun_code/Fun1_7.cpp
@@ -12,10 +12,9 @@
 
 #include <cmath>
 #include "cfloat"
-#include <w32api/sqltypes.h>
 
 int doSomething(int n);
-__int64 fibI( int n);
+long long fibI( int n);
 
 void Fun1_7(unsigned int n){
     for(unsigned int i = 0; i<n; i += 1 + n/2013) {//循环：但迭代至2013次，与n无
@@ -23,7 +22,7 @@ void Fun1_7(unsigned int n){
         if (1 + 1 != 2) goto UNREACHABLE;//分支：条件永非，转向无效
         if(n * n < 0) doSomething(n);//分支：条件永非，调用无效
         if( 2 == (n*n) % 5) Fun1_7(n+1);//分支：条件永非，调用无效
-        __int64 f = fibI(n); if((12 < n) && (sqrt(f) * sqrt(f) == f)) Fun1_7(n-1);//永非
+        long long f = fibI(n); if((12 < n) && (sqrt(f) * sqrt(f) == f)) Fun1_7(n-1);//永非
 
     }
 }
@@ -33,8 +32,8 @@ int doSomething(int n){
 }
 
 
-__int64 fibI( int n){//迭代法：O(n)
-    __int64 g = 0, f = 1;//初始化fib(-1)和fib(0)
+long long fibI( int n){//迭代法：O(n)
+    long long g = 0, f = 1;//初始化fib(-1)和fib(0)
     while (0 < n--){
         g = g + f;
         f = g - f;
